refactor: factor out local_store
